userservice/requirements.txt
Django
django-cors-headers
django-rest-swagger
djangorestframework
djangorestframework-jwt
PyJWT
PyYAML
wheel
psycopg2
requests
